refactor(cli): extract analyze_flow bulk review actions
Moves the three bulk review-action functions (_bulk_accept_rows,
_bulk_skip_rows, _bulk_apply_rows) out of
amx/cli_support/commands/analyze_flow.py into a focused
amx/cli_support/_analyze_bulk_actions.py.

Each function operates on the pending-review queue + history store;
no per-row UI interaction. analyze_flow.py re-exports the three names
so any caller importing the underscore form keeps working.

analyze_flow.py: 2094 -> 1990 LOC (-104).
